put tasks handler into different go module
